24 hour version of our game
This game is a 3D bullet hell that allows you to change from first-person, third-person, and top-down camera views. The goal is to eliminate all of the enemies in order to break the sheild to kill the final boss. This specific version is what we were able to come up with in 24 hours at codeday.