Balancing, lag reduction, and cleaning [Not in 24 hour codeday]
This version is not all made in the 24 hours of codeday and was made within an hour after the next day. This version nerfs and buffs some enemy units, removes duplicate units, nerfs player bullet range to 40 units (to remove sniping exploit), reduces rotation by half to improve aiming of player, and this version makes enemy bullets despawn after 20 seconds to reduce lag. This version also cleans code to fit more of my basic standards and removes commented code that is not needed.